Appointments/requirements.txt
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.1.0
Django==3.2.13
django-cors-headers==3.13.0
django-crontab==0.7.1
django-filter==22.1
djangorestframework==3.13.1
idna==3.3
importlib-metadata==4.12.0
install==1.3.5
Markdown==3.3.7
Pillow==9.2.0
PyMySQL==1.0.2
pytz==2022.1
requests==2.28.1
sqlparse==0.4.2
urllib3==1.26.10
zipp==3.8.0
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.1.0
Django==3.2.13
django-cors-headers==3.13.0
django-crontab==0.7.1
django-filter==22.1
djangorestframework==3.13.1
idna==3.3
importlib-metadata==4.12.0
install==1.3.5
Markdown==3.3.7
Pillow==9.2.0
PyMySQL==1.0.2
pytz==2022.1
requests==2.28.1
sqlparse==0.4.2
urllib3==1.26.10
zipp==3.8.0
